feat: motion cleanup
