+++ src/cala/nodes/prep/__init__.py
@@ -3,14 +3,14 @@ from .denoise import Restore, blur
 from .flatten import butter
 from .glow_removal import GlowRemover
 from .lines import remove_freq, remove_mean
-from .motion import Stabilizer
+from .motion import Anchor
 from .r_estimate import SizeEst
 
 __all__ = [
     "blur",
     "GlowRemover",
     "remove_background",
-    "Stabilizer",
+    "Anchor",
     "SizeEst",
     "butter",
     "remove_mean",
